New function: getAperturePixelCountImage(self, firstSec, integrationTime, y_values, x_values, y_sky, x_sky, apertureMask, skyMask, ...)
y_values and x_values are the (x,y) position for all pixels that are enclosed (including partially enclosed) in the first aperture.
apertureMask is the mask for the first aperture (smaller radius, for the object).
y_sky and y_sky are the (x,y) position for all pixels that are enclosed in the sky aperture (with radius2).
skyMask is the mask for the sky aperture.

This function results in {'image', 'effIntTimes', 'SkyCountSubtractedPerPixel', 'lightcurve'}

'SkyCountSubtractedPerPixel is the average sky background per pixel in the sky aperture (using pixels between radius1 and radius2).
'lightcurve' is the total number of photons in the first aperture (object) with sky subtracted and partial pixels taken into account.